ocr/requirements.txt
requests
pytesseract
Pillow
pillow-heif
lxml
python-dotenv
